mod en SintomasController

+++ src/main/java/com/example/sintomedic/repositorios/SintomasRepositorio.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SintomasRepositorio extends JpaRepository<Sintoma, Integer> {
+public interface SintomasRepositorio extends JpaRepository<Sintoma, Long> {
 
 
 }
